ticket_booking/requirements.txt
amqp==5.3.1
APScheduler==3.11.1
asgiref==3.8.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.10.5
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-rest-auth==7.0.1
Django==4.2.20
django-appconf==1.1.0
django-background-tasks==1.2.8
django-celery-email==3.0.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.2
msgpack==1.1.0
numpy==2.3.4
pandas==2.3.3
pillow==10.4.0
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
redis==5.2.1
reportlab==4.4.4
requests==2.32.5
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
